Remove one DB implementation for tests as there is already one
